Add files via upload
RetroMaze was created on WebSim, mostly with Claude Sonnet 3.5, using music from FolkRNN.org (An RNN trained on a dataset from TheSession.org's library of Public Domain traditional music notated in ABC music notation) for the background music. It is a maze where you can set waypoints to jump back to in the current level. You can set the waypoint in one level and it remains set in the same location on the subsequent levels, giving you a way to jump from a location in the maze to another when enemies get too close. You collect little yellow dots and try not to get shot with "arrows" or caught by colored triangles. The mazes are multicursal with looping paths. It has a retro-like computer styling, and uses oscillators for music and sound effects instead of WAV, MIDI, or MP3. For mobile tablets with touch,  as well as computers with keyboards.