Add files via upload
SpacePlane is what I am calling this one for now. It's a Star Fox-like game with just one level currently. The player must avoid running into obstacles, and fly through rings (toruses). There are shield and invincibility powerups. There are also fuel powerups that are not implemented in this version. There is a working shield system, working radar, and hull integrity percentage. There is also scoring, bonuses, and a 3D particle starfield. The music has not been added yet, but there are oscillator sound effects. I think of it as a paper airplane flying through space. Fun to control, too. Don't try getting every ring, it is not possible.