Add files via upload
A game where you try to gain the most cells in the grid by selecting neighboring colors each turn. Includes computer players, and the ability to play human vs. human on the same device. A simple concept, but somewhat challenging at first. The game works for Mobile or computer. The sounds are a little glitchy, but it works okay.